# apps/data-processing/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sentiment Analysis API",
    "description": "Exposes sentiment analysis for Node.js backend integration",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/account-operations/ingest": {
      "post": {
        "tags": [
          "account-operations"
        ],
        "summary": "Trigger Ingestion",
        "description": "Trigger ingestion of account operations.\n\nAdmin-only endpoint for backfill and incremental ingestion.\n\nArgs:\n    request: Ingestion parameters\n    \nReturns:\n    Ingestion statistics",
        "operationId": "trigger_ingestion_api_account_operations_ingest_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/account-operations/status": {
      "get": {
        "tags": [
          "account-operations"
        ],
        "summary": "Get Status",
        "description": "Get ingestion status for an account.\n\nArgs:\n    account_id: Optional account ID\n    \nReturns:\n    Cursor status",
        "operationId": "get_status_api_account_operations_status_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Account ID",
              "title": "Account Id"
            },
            "description": "Account ID"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/account-operations/reset-cursor": {
      "post": {
        "tags": [
          "account-operations"
        ],
        "summary": "Reset Cursor",
        "description": "Reset the ingestion cursor for an account.\n\nWarning: This will cause re-ingestion of operations.\n\nArgs:\n    request: Reset parameters\n    \nReturns:\n    Success status",
        "operationId": "reset_cursor_api_account_operations_reset_cursor_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetCursorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Reset Cursor Api Account Operations Reset Cursor Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/account-operations/operations": {
      "get": {
        "tags": [
          "account-operations"
        ],
        "summary": "Get Operations",
        "description": "Get account operations from the database.\n\nArgs:\n    account_id: Account ID\n    limit: Maximum number of operations\n    from_ledger: Starting ledger\n    to_ledger: Ending ledger\n    \nReturns:\n    List of operations",
        "operationId": "get_operations_api_account_operations_operations_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Account ID",
              "title": "Account Id"
            },
            "description": "Account ID"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum operations",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum operations"
          },
          {
            "name": "from_ledger",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Starting ledger",
              "title": "From Ledger"
            },
            "description": "Starting ledger"
          },
          {
            "name": "to_ledger",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Ending ledger",
              "title": "To Ledger"
            },
            "description": "Ending ledger"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OperationResponse"
                  },
                  "title": "Response Get Operations Api Account Operations Operations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/news": {
      "get": {
        "summary": "Get News",
        "description": "Return recent articles with optional asset and entity filters.",
        "operationId": "get_news_news_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 168,
              "minimum": 1,
              "default": 24,
              "title": "Hours"
            }
          },
          {
            "name": "asset",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional primary asset code filter",
              "title": "Asset"
            },
            "description": "Optional primary asset code filter"
          },
          {
            "name": "entity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional detected entity filter (example: Soroban)",
              "title": "Entity"
            },
            "description": "Optional detected entity filter (example: Soroban)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NewsArticleResponse"
                  },
                  "title": "Response Get News News Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CursorStatusResponse": {
        "properties": {
          "stream_id": {
            "type": "string",
            "title": "Stream Id",
            "description": "Stream ID"
          },
          "last_ingested_ledger": {
            "type": "integer",
            "title": "Last Ingested Ledger",
            "description": "Last ingested ledger"
          },
          "safe_ledger": {
            "type": "integer",
            "title": "Safe Ledger",
            "description": "Safe ledger for rollback"
          },
          "last_event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Event Id",
            "description": "Last event ID"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Status (idle, ingesting, failed)"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message",
            "description": "Error message if failed"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At",
            "description": "Last update time"
          }
        },
        "type": "object",
        "required": [
          "stream_id",
          "last_ingested_ledger",
          "safe_ledger",
          "status"
        ],
        "title": "CursorStatusResponse",
        "description": "Response model for cursor status."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestionRequest": {
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id",
            "description": "Account ID to ingest operations for"
          },
          "from_ledger": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Ledger",
            "description": "Starting ledger for backfill"
          },
          "to_ledger": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Ledger",
            "description": "Ending ledger for backfill"
          },
          "max_operations": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Operations",
            "description": "Maximum operations to ingest"
          }
        },
        "type": "object",
        "title": "IngestionRequest",
        "description": "Request model for triggering ingestion."
      },
      "IngestionResponse": {
        "properties": {
          "operations_processed": {
            "type": "integer",
            "title": "Operations Processed",
            "description": "Total operations processed"
          },
          "operations_ingested": {
            "type": "integer",
            "title": "Operations Ingested",
            "description": "New operations ingested"
          },
          "operations_duplicate": {
            "type": "integer",
            "title": "Operations Duplicate",
            "description": "Duplicate operations skipped"
          },
          "operations_failed": {
            "type": "integer",
            "title": "Operations Failed",
            "description": "Failed operations"
          },
          "start_ledger": {
            "type": "integer",
            "title": "Start Ledger",
            "description": "Starting ledger"
          },
          "end_ledger": {
            "type": "integer",
            "title": "End Ledger",
            "description": "Ending ledger"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time",
            "description": "Start time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time",
            "description": "End time"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds",
            "description": "Duration in seconds"
          }
        },
        "type": "object",
        "required": [
          "operations_processed",
          "operations_ingested",
          "operations_duplicate",
          "operations_failed",
          "start_ledger",
          "end_ledger"
        ],
        "title": "IngestionResponse",
        "description": "Response model for ingestion results."
      },
...
      "NewsArticleResponse": {
        "properties": {
          "article_id": {
            "type": "string",
            "title": "Article Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "primary_asset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Asset"
          },
          "asset_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Asset Codes",
            "default": []
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Categories",
            "default": []
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords",
            "default": []
          },
          "detected_entities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Detected Entities",
            "default": []
          },
          "onchain_entity_links": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Onchain Entity Links",
            "default": []
          },
          "sentiment_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment Score"
          },
          "sentiment_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment Label"
          },
          "indicator": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SentimentIndicatorResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "article_id",
          "title"
        ],
        "title": "NewsArticleResponse"
      },
      "OperationResponse": {
        "properties": {
          "operation_id": {
            "type": "string",
            "title": "Operation Id",
            "description": "Operation ID"
          },
          "tx_id": {
            "type": "string",
            "title": "Tx Id",
            "description": "Transaction ID"
          },
          "operation_type": {
            "type": "string",
            "title": "Operation Type",
            "description": "Type of operation"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount",
            "description": "Amount"
          },
          "asset_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Code",
            "description": "Asset code"
          },
          "asset_issuer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Issuer",
            "description": "Asset issuer"
          },
          "to_account": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Account",
            "description": "Destination account"
          },
          "from_account": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Account",
            "description": "Source account"
          },
          "ledger": {
            "type": "integer",
            "title": "Ledger",
            "description": "Ledger sequence"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At",
            "description": "Creation time"
          }
        },
        "type": "object",
        "required": [
          "operation_id",
          "tx_id",
          "operation_type",
          "ledger"
        ],
        "title": "OperationResponse",
        "description": "Response model for account operation."
      },
...
      "ResetCursorRequest": {
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id",
            "description": "Account ID"
          },
          "ledger": {
            "type": "integer",
            "title": "Ledger",
            "description": "Ledger to reset to",
            "default": 0
          }
        },
        "type": "object",
        "title": "ResetCursorRequest",
        "description": "Request model for resetting cursor."
      },
...
      "SentimentIndicatorResponse": {
        "properties": {
          "score": {
            "type": "number",
            "title": "Score"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "hex_color": {
            "type": "string",
            "title": "Hex Color"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "display_text": {
            "type": "string",
            "title": "Display Text"
          }
        },
        "type": "object",
        "required": [
          "score",
          "color",
          "hex_color",
          "label",
          "display_text"
        ],
        "title": "SentimentIndicatorResponse",
        "description": "Visual indicator fields attached to every sentiment-bearing response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}